created health bar component
created a new health bar component, which will be used my the monster
class that will be created. this new component has the logic from the
battle scene for creating the health bar container, and contains new
game logic for creating an animation for when a monster takes damage.

added in new images for the health bar shadow, which is shown when the
monster takes damage.

new assets can be found in the assets tag under the health-bar-component.zip

Signed-off-by: Scott Westover <scottwestover2006@gmail.com>