[GH-37] fixed bug tied to unhandled player input in battle scene
Added additional safety guards to handle player input issues in the
battle scene at the start of the battle were the player could start
pressing the okay key in the pre-battle-info state, which would cause
multiple lines of text to be rendered, since we don't wait for the animation
to complete.

To address the issue, I locked the controls at the start of the scene and
only unlock them once the state machine gets to the pre battle info state
and the animations are done.

GitHub Issue: https://github.com/devshareacademy/monster-tamer/issues/37

Signed-off-by: Scott Westover <scottwestover2006@gmail.com>