[GH-61] fixed bug with opening menu when dialog on screen
Addressed the bug in the world scene were the player could open the main
menu while the dialog menu was on the screen. Issue was addressed by
adding a safe guard check for when the player presses the `enter` key
to return early if the dialog ui is open.

GitHub Issue: https://github.com/devshareacademy/monster-tamer/issues/61

Signed-off-by: Scott Westover <scottwestover2006@gmail.com>