[GH-63] fixed bug with saving player direction
Addressed bug were the players direction would not be saved properly if
the player never actually moved in the game. To address this issue, I
added a new callback to be invoked when the characters sprite direction
has changed. This callback is optional, and when invoked, the new direction
will be stored in the data manager.

GitHub Issue: https://github.com/devshareacademy/monster-tamer/issues/63

Signed-off-by: Scott Westover <scottwestover2006@gmail.com>