[GH-101] created common base scene
created a reusable base scene class that will be extended
by all of the scenes in our game. this scene has the common logic for logging
and controls, and in the future will have additional logic for making
the game full screen and cleaning up events when the scene is shutdown.

Signed-off-by: Scott Westover <scottwestover2006@gmail.com>