[GH-98] refactored npc message handling to support events
Refactored the creation of npcs to no longer look for the frame and
messages fields on the data in the Tiled JSON data file. Instead, there
is a new `npcs.json` file that will be used for this data and more.

The new json file includes an array of npc data, where the key is a unique
id for the npc. The data in the Tiled JSON data has been updated to include
this new field, while the old fields have been cleaned up. The new json
file includes the npc frame id for the asset that will be used and it
includes a new `events` field which is an array of object that will be
used for the npc when the player interacts with them.

The new data structure will allow us to do things like have the npc talk
to the player, heal the player, give items, battle, etc. To support this
the npc class and world scene classes have been updated to pass this
data when the npc is created. Then, when the player interacts with the npc
a new method in the world scene is called. This method keeps track of which
events have been processed, and will take action on the relevant event
data.

As an example, if an npc has 2 events that are just messages, then the player
will see the dialog ui (the same as before), and after talking to the npc
twice, the dialog ui goes away (same as before).

Finally, started adding in logic for each of the custom event types, like
healing the player and fading the scene in and out.

Signed-off-by: Scott Westover <scottwestover2006@gmail.com>