[GH-98] enhanced world scene and battle scene to spawn player when KO
updated the world scene data to support a new property for when the player
is knocked out during a battle. when this is set to `true`, the world
scene will attempt to lookup the players re-spawn location based on the
data in the TILED JSON data file.

Once the location is found, the world scene will update the data in the
data manager to match that location and player direction.

Signed-off-by: Scott Westover <scottwestover2006@gmail.com>