[GH-115] updated battle scene to have dynamic monster data
Updated the battle scene to expect data to be passed when the scene is
initialized. The data now expects two arrays, one for the monsters in the
enemy party and one for the monsters in the player party.

Refactored all data related to monsters to now live in a JSON file.
Updated the battle scene, world scene, and data manager to use a new
util method for getting monster data from that JSON file.

Enhanced the battle scene to now updated the scene data and data manager
data of the monsters between each round in the battle.

Signed-off-by: Scott Westover <scottwestover2006@gmail.com>