[GH-119] added support for items on world map
Updated world scene to pull items from the level data
and to created objects in the phaser scene. Created a
new Item class to represent these objects.

Updated the data manager and data utils to allow
adding new items to the existing inventory and to
update their quantity. Also, added new data method
for persisting the items that are picked up from the
world scene, so we only show items that haven't been
picked up when the scene reloads.

Updated the player class to allow a new array of
objects to be passed that can be used for checking
for collisions.

Signed-off-by: Scott Westover <scottwestover2006@gmail.com>