[GH-117] added support for entering another area
Updated the player class to support having a transition
layer that detects if the player needs to move to another
level.

Updated the player config to support the new layer and a callback
to know when the player character enters that area. Updated the
world scene to listen for the callback and reload the
world scene with the next area data.

Commit with test world scene transition: https://github.com/devshareacademy/monster-tamer/commit/c00ea6116b09746906413640e075d56e3d10aca6

Signed-off-by: Scott Westover <scottwestover2006@gmail.com>