[GH-117] refactored npc dialog logic
Refactored logic for npcs to no longer have the messages
in the Tiled JSON file, and to have a new dedicated JSON file
with npc character data. Refactored the data structure
to now be in a object structure which will support a variety of event
types.

This will be the structure to support the npcs performing various
actions in the game like: dialog, give items, trade, battle, etc.
Added support to data utils to load and parse the new structure.

Signed-off-by: Scott Westover <scottwestover2006@gmail.com>