[GH-138] added support for multiple npc animations
Updated the npc class to support a custom animation
prefix in order to allow multiple npcs to be animated
and moving around the game.

Added new npc animations to the animations and npc
json files. Updated the npc json file to support
a new animation prefix field that is passed to the
npc class when creating a new npc instance.

Signed-off-by: Scott Westover <scottwestover2006@gmail.com>