[GH-138] added event zones to world scene
Updated map data from Tiled to include a new "Events"
layer, which is used for creating zone game objects
in the world scene.

These zone game objects will be used for checking
overlaps with the player, and when there is an overlap
this will be the trigger for our cutscene or in
game event.

This change just creates the zone game object.

Added support for rending out the zone game objects
by creating a rectangle game object. The creation
of these game objects are wrapped behind a config
variable that can be turned on/off from the config
file.

Signed-off-by: Scott Westover <scottwestover2006@gmail.com>