[GH-138] added event data structure
Created new events json file that has the core metadata
needed for an in game event. Created data types for
the various event types. Updated preloader to load in
the new json file, and enhanced the data utils class
to support getting an event from the cache by the
id of the event.

Signed-off-by: Scott Westover <scottwestover2006@gmail.com>