GH-146 added logic to spawn player from level data on new game
added logic to support loading a new object layer from tiled that has the player spawn location. the new field makes it easier to spawn the player into the world when the map data changes. this was needed since we resized the original map data, which broke the hard coded spawn location in our data manager.

Signed-off-by: Scott Westover <scottwestover2006@gmail.com>