GH-146 fixed bug with npc and events spawning after level updates
had to update the npc and events json files for when we spawn objects and npcs during events. this was needed after the level was updated to have a new size, which caused the original locations to no longer be valid. this was similar to the fix that was needed for the player spawn location.

Signed-off-by: Scott Westover <scottwestover2006@gmail.com>